requirements.txt
# Existing requirements from Week 1 & 2
#Django>=4.2.0
#psycopg2-binary>=2.9.0
#python-decouple>=3.8

# New API-specific requirements
djangorestframework>=3.14.0
djangorestframework-gis>=1.0.0
django-filter>=23.0
django-cors-headers>=4.3.0
drf-spectacular>=0.26.0  # API documentation
django-extensions>=3.2.0
requests>=2.31.0  # For testing external API calls

# Testing and development
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0